requirements.txt
typing
pandas

google-api-core
google-cloud-documentai